requirements.txt
astor==0.8.1
attrs==22.2.0
bandit==1.7.4
certifi==2022.12.7
charset-normalizer==2.1.1
darglint==1.8.1
docutils==0.19
eradicate==2.1.0
Faker==15.3.4
flake8==3.9.2
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.12.6
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
gitdb==4.0.10
GitPython==3.1.30
hvac==1.0.2
idna==3.4
isort==5.11.4
mccabe==0.6.1
pbr==5.11.0
pep8-naming==0.13.2
psycopg2-binary==2.9.5
pydantic==1.10.3
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.13.0
pyhcl==0.4.4
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
restructuredtext-lint==1.4.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==4.1.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
wemake-python-styleguide==0.17.0
django-split-settings==1.2.0
django-debug-toolbar==3.8.1
redis==4.4.0
django-redis==5.2.0
hiredis==2.1.0